+++ CHANGES.rst
@@ -23,6 +23,8 @@ Improvements
   :pr:`6918`, thanks :user:`duartegalvao, unconventionaldotdev`)
 - Add option in the invitation form to lock registrations to the specified email address
   (:issue:`6803`, :pr:`6972`, thanks :user:`duartegalvao, unconventionaldotdev`)
+- Add plugin support for scanning custom QR codes in the Check-in app (:pr:`6954`,
+  thanks :user:`SegiNyn`)
 
 Bugfixes
 ^^^^^^^^